+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/ai-assistant-plugin-sidebar/index.tsx
@@ -4,7 +4,10 @@
 import { JetpackEditorPanelLogo } from '@automattic/jetpack-shared-extension-utils';
 import { useAnalytics } from '@automattic/jetpack-shared-extension-utils';
 import { PanelBody, PanelRow, BaseControl } from '@wordpress/components';
+import { store as coreStore } from '@wordpress/core-data';
+import { useSelect } from '@wordpress/data';
 import { PluginPrePublishPanel, PluginDocumentSettingPanel } from '@wordpress/edit-post';
+import { store as editorStore } from '@wordpress/editor';
 import { __ } from '@wordpress/i18n';
 import debugFactory from 'debug';
 import React from 'react';
@@ -19,12 +22,14 @@ import Proofread from '../proofread';
 import TitleOptimization from '../title-optimization';
 import UsagePanel from '../usage-panel';
 import {
+	CoreSelect,
 	JetpackSettingsContentProps,
 	PLACEMENT_DOCUMENT_SETTINGS,
 	PLACEMENT_JETPACK_SIDEBAR,
 	PLACEMENT_PRE_PUBLISH,
 } from './types';
 import Upgrade from './upgrade';
+import type * as EditorSelectors from '@wordpress/editor/store/selectors';
 
 import './style.scss';
 
@@ -89,6 +94,21 @@ export default function AiAssistantPluginSidebar() {
 	const { checkoutUrl } = useAICheckout();
 
 	const { tracks } = useAnalytics();
+
+	const isViewable = useSelect( select => {
+		const postTypeName = ( select( editorStore ) as typeof EditorSelectors ).getCurrentPostType();
+		// The coreStore select type lacks the getPostType method, so we need to cast it to the correct type
+		const postTypeObject = ( select( coreStore ) as unknown as CoreSelect ).getPostType(
+			postTypeName
+		);
+
+		return postTypeObject?.viewable;
+	}, [] );
+	// If the post type is not viewable, do not render my plugin.
+	if ( ! isViewable ) {
+		return null;
+	}
+
 	const title = __( 'AI Assistant', 'jetpack' );
 
 	const panelToggleTracker = placement => {