+++ tools/e2e-commons/pages/wp-admin/jetpack-dashboard.js
@@ -1,6 +1,6 @@
-import WpPage from '../wp-page.js';
-import logger from '../../logger.js';
 import { resolveSiteUrl } from '../../helpers/utils-helper.js';
+import logger from '../../logger.js';
+import WpPage from '../wp-page.js';
 
 export default class JetpackDashboardPage extends WpPage {
 	constructor( page ) {
@@ -14,6 +14,8 @@ export default class JetpackDashboardPage extends WpPage {
 	 * Determine if the site is connected to WordPress.com,
 	 * based on the visibility of a connection card element and the text in this element
 	 * Should be used to assert a site is connected.
+	 *
+	 * @returns {boolean} If the site is connected.
 	 */
 	async isSiteConnected() {
 		logger.step( 'Checking that site is connected' );
@@ -27,6 +29,8 @@ export default class JetpackDashboardPage extends WpPage {
 	 * Determine if a WordPress.com is connected or not,
 	 * based on the visibility of a second connection card element and the text in this element
 	 * Should be used to assert a user is connected to WordPress.com. To check a user is not connected, see isUserNotConnected()
+	 *
+	 * @returns {boolean} If a user is connected.
 	 */
 	async isUserConnected() {
 		logger.step( 'Checking that WordPress.com user is connected' );
@@ -38,6 +42,8 @@ export default class JetpackDashboardPage extends WpPage {
 	 * Determine if a WordPress.com is connected or not,
 	 * based on the visibility of a second connection card element and the text in this element
 	 * Should be used to assert a user is NOT connected to WordPress.com. To check a user is connected, see isUserConnected()
+	 *
+	 * @returns {boolean} If a user is not connected.
 	 */
 	async isNotUserConnected() {
 		logger.step( 'Checking that WordPress.com user is not connected' );