+++ projects/js-packages/publicize-components/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.54.5] - 2024-06-24
+### Added
+- Added Threads to the social connections services [#37977]
+
+### Fixed
+- Fixed skipped connections id [#37810]
+- Social | Replaced + icon with connections management link in the post editor [#37961]
+
 ## [0.54.4] - 2024-06-17
 ### Fixed
 - Social: Fixed broken connections reconnect link to point it to new connections UI [#37869]
@@ -741,6 +749,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated package dependencies. [#24470]
 
+[0.54.5]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.54.4...v0.54.5
 [0.54.4]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.54.3...v0.54.4
 [0.54.3]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.54.2...v0.54.3
 [0.54.2]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.54.1...v0.54.2